[UDG] fix ordering of the level sets
When using the level set images, we currently assume that the entries are
ordered according to their indices in grid index set. As we are using
YaspGrid for the structured grid, this means column major or Fortran
style ordering. We add a check to make sure that the numpy arrays passed
to duneuro-py conform to this ordering.